requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.7
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
packaging==23.2
Pillow==10.1.0
psycopg==3.1.12
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.12.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
